+++ lib/hosts.js
@@ -1,0 +1,4 @@
+export default {
+  develop: 'preview.contentful.com',
+  production: 'cdn.contentful.com'
+}
